+++ molten-test/src/test/java/com/hotels/molten/test/JUnitJupiterTest.java
@@ -20,10 +20,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 
 @Slf4j
-public class JunitTest {
+public class JUnitJupiterTest {
     @Test
     public void should() {
-        LOG.info("JUnit tests are running.");
+        LOG.info("JUnit 5 tests are running.");
     }
 
 }