+++ CHANGELOG.md
@@ -6,5 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0]
+### Added
+- Added spring-boot 2 auto-configuration support with `molten-spring-boot` module.
+- Revamped mockito support with auto-configuration. `@Mock` now supports reactor types.
+
+### Deprecated
+- `@ReactiveMock` and related classes are now deprecated.
+
 ## [1.0.0]
 - Initial release